refactor(test): move fake api-server tests in separate file
also, remove all calls to `logger.SetLevel(logr.DebugLevel)`
